dont recreate files and directories on rerun

+++ stubs/runtime-with-vendor-download.php
@@ -4,7 +4,9 @@ ini_set('display_errors', '1');
 
 error_reporting(E_ALL);
 
-mkdir('/tmp/opcache');
+if (! file_exists('/tmp/opcache')) {
+    mkdir('/tmp/opcache');
+}
 
 $appRoot = $_ENV['LAMBDA_TASK_ROOT'];
 
@@ -19,16 +21,18 @@ $appRoot = $_ENV['LAMBDA_TASK_ROOT'];
 |
 */
 
-fwrite(STDERR, 'Downloading the application vendor archive...');
+if (! file_exists('/tmp/vendor')) {
+    fwrite(STDERR, 'Downloading the application vendor archive...');
 
-exec(sprintf('/opt/awscli/aws s3 cp s3://%s/%s-vendor.zip /tmp/vendor.zip',
-    $_ENV['VAPOR_ARTIFACT_BUCKET_NAME'],
-    $_ENV['VAPOR_ARTIFACT_NAME']
-));
+    exec(sprintf('/opt/awscli/aws s3 cp s3://%s/%s-vendor.zip /tmp/vendor.zip',
+        $_ENV['VAPOR_ARTIFACT_BUCKET_NAME'],
+        $_ENV['VAPOR_ARTIFACT_NAME']
+    ));
 
-exec('unzip /tmp/vendor.zip -d /tmp/vendor');
+    exec('unzip /tmp/vendor.zip -d /tmp/vendor');
 
-unlink('/tmp/vendor.zip');
+    unlink('/tmp/vendor.zip');
+}
 
 /*
 |--------------------------------------------------------------------------
